# src/v2/openapi.yaml
openapi: 3.0.1
info:
  title: Backbone API
  version: "2.0"
servers:
  - url: /api/v2

paths:
  /recommender:
    post:
      tags:
      - Scan Services
      summary: |
        Generates companion recommendation with the help of Ecosystem
        specific recommendation engine and writes it into RDS.
      parameters:
      - name: persist
        in: query
        schema:
          type: boolean
          default: true
        description: |
          Whether to persist generated report into RDS.
      - name: check_license
        in: query
        schema:
          type: boolean
          default: false
        description: |
          Whether to do license analyses for the recommend packages.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecommenderRequest'
        required: true
      responses:
        200:
          description: Stack report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StackRecommendationResult'
        400:
          description: Bad request
          content: {}

  /stack_aggregator:
    post:
      tags:
      - Scan Services
      summary: |
        Aggregates vulnerabilities, licenses and github statistics associate with list of given packages and writes it into RDS.
      parameters:
      - name: persist
        in: query
        schema:
          type: boolean
          default: true
        description: |
          Whether to persist generated report into RDS.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StackAggregatorRequest'
        required: true
      responses:
        200:
          description: Stack report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StackAggregatorResult'
        400:
          description: Bad request
          content: {}

components:
  schemas:
    RecommenderRequest:
      title: RecommenderRequest
      allOf:
        - $ref: '#/components/schemas/StackAggregatorRequest'
      description: EPV details for stack report generation

    StackAggregatorRequest:
      title: StackAggregatorRequest
      type: object
      required:
        - registration_status
        - uuid
        - external_request_id
        - ecosystem
        - packages
        - manifest_name
        - manifest_file_path
      properties:
        registration_status:
          type: string
          example: FREETIER
        uuid:
          type: string
          format: uuid
        external_request_id:
          type: string
          example: 229e7da76708fe374d8c10fa752e72989f
        show_transitive:
          type: boolean
          default: true
          description: |
            This is required to enable or disable the transitive support
        ecosystem:
          $ref: '#/components/schemas/Ecosystem'
        manifest_name:
          type: string
        manifest_file_path:
          type: string
        packages:
          type: array
          items:
            $ref: '#/components/schemas/Package'
      description: EPV details for stack report generation

    Ecosystem:
      title: Ecosystem
      type: string
      enum:
        - maven
        - pypi
        - npm
      description: List of supported package ecosystems

    VulnerabilityFields:
      title: VulnerabilityFields
      required:
        - cvss
        - cvss_v3
        - severity
        - title
        - id
        - url
      properties:
        cve_ids:
          type: array
          items:
            type: string
          example:
            - CVE-2014-0474
            - CVE-2014-0475
        cvss:
          type: number
          format: float
          example: 9.8
        cwes:
          type: array
          items:
            type: string
          example:
            - CWE-185
            - CWE-730
        cvss_v3:
          type: string
          example: "CVSS:3.0/AV:N/AC:L/PR:N/UI:N/S:U/C:H/I:H/A:H"
        severity:
          type: string
          enum:
            - low
            - medium
            - high
            - critical
        title:
          type: string
          example: SQL Injection
        id:
          type: string
          example: SNYK-PYTHON-DJANGO-40027
        url:
          type: string
          example: https://snyk.io/vuln/SNYK-PYTHON-DJANGO-4002
        malicious:
          type: boolean
          default: true
        patch_exists:
          type: boolean
          default: false
          example: true
        fixable:
          type: boolean
          default: false
          example: true
        exploit:
          type: string
          enum:
            - High
            - Functional
            - Proof of Concept
            - Unproven
            - Not Defined
          example: High
        description:
          type: string
          example: |
            "## Overview\r\n[`django`](https://pypi.python.org/pypi/django) is a high-level Python Web framework that encourages rapid development and clean, pragmatic design.\r\n\r\nAffected versions of this package expose sensitive information due to not properly restricting the use of a query string that performs certain object filtering. An attacker may obtain sensitive information via a series of requests containing regular expressions, as demonstrated by a `created_by__password__regex` parameter.\r\n\r\n## References\r\n- [Django Vulnerability Description](http://www.djangoproject.com/weblog/2010/dec/22/security/)\r\n- [GitHub Commit](https://github.com/django/django/commit/732198ed5c)\r\n- [Redhat Bugzilla](https://bugzilla.redhat.com/show_bug.cgi?id=665373)\r\n- [Openwall](http://www.openwall.com/lists/oss-security/2011/01/03/5)\r\n- [CVE](https://web.nvd.nist.gov/view/vuln/detail?vulnId=CVE-2010-4534)\r\n"
        fixed_in:
          type: array
          items:
            type: string
          example:
            - 2.6.7
            - 2.7.5
            - 2.8.0
        references:
          type: array
          items:
            type: object
            properties:
              title:
                type: string
                example: Redhat Security Advisory
              url:
                type: string
                example: https://access.redhat.com/security/cve/CVE-2013-2186
          description: |
            All vulnerability details as provided from graph DB

    Package:
      title: Package
      type: object
      required:
        - name
        - version
      properties:
        name:
          type: string
          example: django
        version:
          type: string
          example: 1.2.1
        dependencies:
          type: array
          items:
            $ref: '#/components/schemas/Package'
          example:
            - requests:
                name: requests
                version: 1.2.3
            - six:
                name: six
                version: 5.1.1
      description: Package details

    UnknownLicenses:
      title: UnknownLicenses
      type: object
      properties:
        component_conflict:
          type: array
          items:
            type: object
            properties:
              package:
                type: string
                example: django
              license1:
                type: string
                example: BSD License (BSD)
              license2:
                type: string
                example: General Public License V3(GPL)
          description: |
            List of conflicting licenses
        unknown:
          type: array
          items:
            type: object
            properties:
              package:
                type: string
                example: django
              license:
                type: string
                example: BSD License (BSD)
            description: |
              List of licenses which are unknown to the platform.
      description: Licesens which are unknown to the platform.

    ConflictPackages:
      title: ConflictPackages
      type: object
      required:
        - package1
        - package2
        - license1
        - license2
      properties:
        package1:
          type: string
          example: django
        license1:
          type: string
          example: BSD License (BSD)
        package2:
          type: string
          example: rustdoc
        license2:
          type: string
          example: General Public License V3(GPL)
      description: Pair of packages which has conflict in license

    LicenseAnalysis:
      title: LicenseAnalysis
      type: object
      required:
        - outlier_packages
        - conflict_packages
        - current_stack_license
        - unknown_licenses
      properties:
        reason:
          type: string
          example: |
            Cannot calculate stack license due to unknown dependencies or license not supported.
        status:
          type: string
          enum:
            - ComponentConflict
            - StackConflict
            - Successful
            - Unknown
            - Failure
        recommended_licenses:
          type: array
          items:
            type: string
        outlier_packages:
          type: array
          items:
            type: object
            properties: {}
        conflict_packages:
          type: array
          items:
            $ref: '#/components/schemas/ConflictPackages'
        current_stack_license:
          type: object
          properties: {}
        unknown_licenses:
          $ref: '#/components/schemas/UnknownLicenses'
        distinct_licenses:
          type: array
          items:
            type: string
      description: License analyses data

    GitHubDetails:
      title: GitHubDetails
      type: object
      properties:
        watchers:
          type: string
          example: "100K"
        total_releases:
          type: string
          example: "12"
        issues:
          type: object
          properties: {}
        pull_requests:
          type: object
          properties: {}
        dependent_repos:
          type: string
        open_issues_count:
          type: string
        latest_release_duration:
          type: string
        forks_count:
          type: string
          example: "123"
        contributors:
          type: string
          example: "12"
        size:
          type: string
        stargazers_count:
          type: string
          example: "50K"
        used_by:
          type: array
          items:
            type: string
        dependent_projects:
          type: string
          example: "11K"
      description: Github details

    PackageDetails:
      title: PackageDetails
      allOf:
        - $ref: '#/components/schemas/Package'
        - type: object
          required:
            - name
            - version
            - latest_version
            - ecosystem
          properties:
            latest_version:
              type: string
              example: 3.0.5
            github:
              $ref: '#/components/schemas/GitHubDetails'
            licenses:
              type: array
              items:
                type: string
              example:
                - Apache License, Version 2.0
            ecosystem:
              $ref: '#/components/schemas/Ecosystem'
            url:
              type: string
              example:
                - https://snyk.io/vuln/pip:django
                - https://snyk.io/vuln/pip:jinja2
          description: Details of the package

    PackageDataWithVulnerabilities:
      title: PackageDataWithVulnerabilities
      allOf:
        - $ref: '#/components/schemas/PackageDetails'
        - type: object
          properties:
            public_vulnerabilities:
              type: array
              description: Publicly known vulnerability details
              items:
                $ref: '#/components/schemas/VulnerabilityFields'
            private_vulnerabilities:
              type: array
              description: |
                Private vulnerability details
              items:
                $ref: '#/components/schemas/VulnerabilityFields'
            recommended_version:
              type: string
              example: 3.0.0
              description: |
                Recommended package version which includes fix for both public and private vulnerabilities.
            vulnerable_dependencies:
              type: array
              items:
                $ref: '#/components/schemas/PackageDataWithVulnerabilities'
              description: |
                List of dependencies which are vulnerable.
          description: Details of the package for the given stack

    RecommendedPackageData:
      title: RecommendedPackageData
      description: Recommended packages data
      allOf:
      - $ref: '#/components/schemas/PackageDetails'
      - type: object
        properties:
          cooccurrence_probability:
            type: number
            format: float
            example: 83.16431
          cooccurrence_count:
            type: string
          topic_list:
            type: array
            items:
              type: string

    StackRecommendationResult:
      title: StackRecommendation
      type: object
      required:
        - _audit
        - uuid
        - external_request_id
        - registration_status
        - recommendation_status
        - companion
        - manifest_file_path
      properties:
        _audit:
          $ref: '#/components/schemas/Audit'
        uuid:
          type: string
          format: uuid
        external_request_id:
          type: string
          example: 229e7da76708fe374d8c10fa752e72989f
        registration_status:
          type: string
          example: REGISTERED
        recommendation_status:
          type: string
          enum:
            - success
            - pgm_error
            - invalid_ecosystem
            - invalid_payload
            - insights_call_failure
          default: success
        companion:
          type: array
          items:
            $ref: '#/components/schemas/RecommendedPackageData'
        manifest_file_path:
          type: string
          example: pom.xml
      description: Application stack recommendations

    Audit:
      title: Audit
      type: object
      required:
        - started_at
        - ended_at
        - api_version
      properties:
        started_at:
          type: string
        ended_at:
          type: string
        api_version:
          type: string

    StackAggregatorResult:
      title: StackAggregatorResult
      type: object
      required:
       - _audit
        - uuid
        - external_request_id
        - registration_status
        - manifest_file_path
        - manifest_name
        - ecosystem
        - unknown_dependencies
        - license_analysis
        - registration_link
        - analyzed_dependencies
      properties:
        _audit:
          $ref: '#/components/schemas/Audit'
        uuid:
          type: string
          format: uuid
        external_request_id:
          type: string
          example: 229e7da76708fe374d8c10fa752e72989f
        registration_status:
          type: string
          example: FREETIER
        manifest_file_path:
          type: string
          example: pom.xml
        manifest_name:
          type: string
          example: dependencies.txt
        ecosystem:
          $ref: '#/components/schemas/Ecosystem'
        unknown_dependencies:
          type: array
          items:
            $ref: '#/components/schemas/Package'
          example:
            - name: rocket-db
              version: 2020.1.alpha
            - name: jsonp
              version: 0.0.1
        license_analysis:
          $ref: '#/components/schemas/LicenseAnalysis'
        registration_link:
          type: string
          format: url
        analyzed_dependencies:
          type: array
          items:
            $ref: '#/components/schemas/PackageDataWithVulnerabilities'
          description: |
            All direct dependencies details regardless of it's vulnerability status
      description: Stack analyses result data